// src/enums/bidder-dashboard.enum.ts
export enum BidderDashboardTableTitlesEnum {
  BIDDER_DASHBOARD_HOME = 'Bidder Dashboard',
}

export enum BidderDashboardColumnEnum {
  ACCOUNT_ID = 'Account Id',
  MARKETPLACE = 'Market Place',
  JOB_ID = 'Job Id',
  LAST_TRIGGER_DATE = 'Last Trigger Date',
  NEXT_TRIGGER_DATE = 'Next Trigger Date',
  REPORT_DATE = 'Report Date',
  BRAND_NAME = 'Brand Name',
  UNIQUE_ID = 'Unique Id',
  STATUS = 'Status',
  AD_TYPE = 'Ad Type',
}

export enum BidderDashboardStatusEnum {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  SCHEDULED = 'SCHEDULED',
  FAILED = 'FAILED',
  COMPLETED = 'COMPLETED',
}

export enum BidderDashboardMarketplaceEnum {
  AMAZON = 'amazon',
  WALMART = 'walmart',
}

export enum BidderDashboardAdTypeEnum {
  SP = 'SP',
  SB = 'SB',
  SD = 'SD',
  SV = 'SV',
}

export enum BidderDashboardTitleEnum {
  BidderDashboardHome = 'Bidder Dashboard',
}